:hammer: refactore route and logic route, implement route register/emial-exist

+++ app/Http/Controllers/API/v1/CronLikesController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\API\v1;
 
 use App\Http\Controllers\Controller;
+use App\Http\Responses\ErrorResponses;
 use App\Http\Responses\SuccessResponses;
 use Illuminate\Http\Request;
 
@@ -15,18 +16,43 @@ class CronLikesController extends Controller
         $this->user = auth()->user();
     }
 
-    public function store(Request $request)
+    public function store($cardId)
     {
-        $cronId = $request->input('cl_fk_cron_id');
 
-        $like = $this->user->likes()->where('cl_fk_cron_id', $cronId)->first();
+        $cron = $this->user->crons()
+            ->where('c_status', 'ACTIVE')
+            ->where('c_id', $cardId)
+            ->first();
+
+        if (!$cron) {
+            return ErrorResponses::notFound(['message' => 'Cron not found']);
+        }
+
+        $like = $this->user->likes()->where('cl_fk_cron_id', $cron->c_id)->first();
 
         if ($like) {
             $like->delete();
             return SuccessResponses::ok([], ['message' => "Like removed successfully"]);
         }
-        $this->user->likes()->create(['cl_fk_cron_id' => $cronId]);
+        $this->user->likes()->create(['cl_fk_cron_id' => $cron->c_id]);
 
         return SuccessResponses::created([], ['message' => "Like created successfully"]);
+    }
+
+    public function getCronLikes($cronId)
+    {
+        $cron = $this->user->crons()
+            ->where('c_status', 'ACTIVE')
+            ->where('c_id', $cronId)
+            ->first();
+
+
+        if (!$cron) {
+            return ErrorResponses::notFound(['message' => 'Cron not found']);
+        }
+
+        $likes = $cron->likes()->get();
+
+        return SuccessResponses::ok(['likes' => $likes], ['message' => 'get likes successfully']);
     }
 }
