+++ app/Http/Controllers/API/v1/UpVoteController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\API\v1;
 
 use App\Http\Controllers\Controller;
+use App\Http\Responses\ErrorResponses;
 use App\Http\Responses\SuccessResponses;
 use Illuminate\Http\Request;
 
@@ -16,18 +17,42 @@ class UpVoteController extends Controller
         $this->user = auth()->user();
     }
 
-    public function store(Request $request)
+    public function store($cronId)
     {
-        $cronId = $request->input('uv_fk_cron_id');
+        $cron = $this->user->crons()
+            ->where('c_status', 'ACTIVE')
+            ->where('c_id', $cronId)
+            ->first();
 
-        $upVote = $this->user->upVotes()->where('uv_fk_cron_id', $cronId)->first();
+        if (!$cron) {
+            return ErrorResponses::notFound(['message' => 'Cron not found']);
+        }
+
+        $upVote = $this->user->upVotes()->where('uv_fk_cron_id', $cron->c_id)->first();
 
         if ($upVote) {
             $upVote->delete();
             return SuccessResponses::ok([], ['message' => "Up vote removed successfully"]);
         }
-        $this->user->upVotes()->create(['uv_fk_cron_id' => $cronId]);
+        $this->user->upVotes()->create(['uv_fk_cron_id' => $cron->c_id]);
 
         return SuccessResponses::created([], ['message' => "Up vote created successfully"]);
+    }
+
+    public function getCronUpVotes($cronId)
+    {
+        $cron = $this->user->crons()
+            ->where('c_status', 'ACTIVE')
+            ->where('c_id', $cronId)
+            ->first();
+
+
+        if (!$cron) {
+            return ErrorResponses::notFound(['message' => 'Cron not found']);
+        }
+
+        $upVotes = $cron->upVotes()->get();
+
+        return SuccessResponses::ok(['up_votes' => $upVotes], ['message' => 'get up votes successfully']);
     }
 }
